:construction_worker:[RUMF-449] tweak build infos
leverage browserstack notion of project and build to group test executions
